Simplify FormDataForm
