+++ frontend/src/pages/RequestorPage/FormDataForm/index.ts
@@ -1,10 +1,6 @@
-export { FormDataForm, FormDataFormRow } from "./FormDataForm";
+export { FormDataForm } from "./FormDataForm";
 export {
   reduceFormDataParameters,
-  createFormDataParameters,
-} from "./data";
-export {
   enforceTerminalDraftParameter as enforceFormDataTerminalDraftParameter,
-  useFormDataForm,
-} from "./hooks";
-export { FormDataParameterSchema, type FormDataParameter } from "./types";
+} from "./data";
+export { FormDataParameterSchema } from "./types";
